+++ GrpB_models/b1_customer_preferences/data_preparation.py
