+++ GrpB_models/b1_customer_preferences/main.py
@@ -17,8 +17,8 @@ from sklearn.preprocessing import MinMaxScaler
 from sklearn.metrics import roc_curve, accuracy_score, f1_score, roc_auc_score, average_precision_score
 from bayes_opt import BayesianOptimization
 
-from datapreparation import import_data, data_cleaning, data_split, label_split
-from hyperparametertuning import get_lgb_opt_params, get_xgb_opt_params
+from data_preparation import import_data, data_cleaning, data_split, label_split
+from hyperparameter_tuning import get_lgb_opt_params, get_xgb_opt_params
 from models import final_classification_model, final_recommendation_model
 
 ####################Data Preparation####################
